+++ src/components/Icons/DownArrow.tsx
@@ -1,10 +1,24 @@
 import React from "react";
-import "../styles/treeView.css";
-import { DownArrowProps } from "../types/componentTypes";
+import "../../styles/treeView.css";
+import { DownArrowProps } from "../../types/componentTypes";
+
+/**
+ * @description DownArrow component that renders a down arrow SVG icon.
+ *
+ * @param {number} [size=10] - The size (height and width) of the arrow icon in pixels. Defaults to 10.
+ * @param {string} [color="#00f0ff"] - The fill color of the arrow icon. Defaults to "#00f0ff" (a light blue).
+ *
+ * @example
+ * <DownArrow size={20} color="red" />
+ *
+ * @example
+ * <DownArrow />
+ */
 
 const DownArrow: React.FC<DownArrowProps> = ({
   size = 10,
   color = "#00f0ff",
+  className = ""
 }) => {
   return (
     <svg
@@ -15,6 +29,7 @@ const DownArrow: React.FC<DownArrowProps> = ({
       id="Layer_1"
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 330 330"
+      className={className}
       // xmlns:xlink="http://www.w3.org/1999/xlink"
       // xml:space="preserve"
     >