+++ src/components/TreeExpander.tsx
@@ -1,16 +1,40 @@
 import React from "react";
 import "../styles/treeView.css";
 import { TreeExpanderProps } from "../types/componentTypes";
-import DownArrow from "./DownArrow";
+import DownArrow from "./Icons/DownArrow";
 import TreeLiner from "./TreeLiner";
 import TreeHeader from "./TreeHeader";
-import TreeText from "./TreeText";
+
+/**
+ * @description TreeExpander component that creates an expandable section in a tree view. This displays a header that, when clicked,
+ * expands or collapses to reveal the content within.  This uses CSS transitions for smooth animation of the height change.
+ *
+ * @param {boolean} [expanded=false] - Controls the initial expanded state of the component. Defaults to `false` (collapsed).
+ * @param {boolean} [isSimpleHeader=true] - Determines whether to render the header using the `TreeText` component (simple) or the `TreeHeader` component (more complex). Defaults to `true` (simple header).
+ * @param {string} header - The text to display in the header of the expandable section.
+ * @param {React.ReactNode} children - The content to be displayed within the expandable section.
+ *
+ * @example
+ * <TreeExpander header="Section 1">
+ *   <p>Content of section 1.</p>
+ * </TreeExpander>
+ *
+ * @example
+ * <TreeExpander header="Advanced Section" expanded isSimpleHeader={false}>
+ *   <p>Content of section 1.</p>
+ * </TreeExpander>
+ */
 
 const TreeExpander: React.FC<TreeExpanderProps> = ({
   expanded: initialExpanded = false,
-  isSimpleHeader = true,
   header,
   children,
+  logo,
+  subHeader,
+  desc,
+  link,
+  sublink,
+  descLink,
 }) => {
   const [isExpanded, setIsExpanded] = React.useState(initialExpanded);
   const [maxHeight, setMaxHeight] = React.useState<number | string>(0);
@@ -38,13 +62,17 @@ const TreeExpander: React.FC<TreeExpanderProps> = ({
         onClick={handleExpander}
       >
         <div className="treeViewExpanderHeader">
-          {isSimpleHeader ? (
-            <TreeText text={header} />
-          ) : (
-            <TreeHeader header={header} />
-          )}
+          <TreeHeader
+            header={header}
+            logo={logo}
+            subHeader={subHeader}
+            desc={desc}
+            link={link}
+            sublink={sublink}
+            descLink={descLink}
+          />
         </div>
-        <DownArrow />
+        <DownArrow className="treeViewExpanderArrow"/>
       </div>
       <div
         className={`treeViewExpanderBodyContainer ${